NTI_Final_Project/requirements.txt
streamlit==1.37.1
torch>=2.2.0
transformers>=4.41.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
pypdf2>=3.0.1
pytesseract>=0.3.10
pillow>=10.3.0
pdf2image>=1.17.0
reportlab>=4.2.0
yake==0.6.0
numpy>=1.26.4
pandas>=2.2.2
rouge-score>=0.1.2
sacrebleu>=2.4.0
